requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.8
django-bootstrap-pagination==1.7.1
django-cors-headers==3.4.0
django-extensions==3.0.4
django-pure-pagination==0.3.0
filter-and-pagination==1.1.14
freezegun==1.0.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
lxml==4.5.2
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.0
parsel==1.6.0
Pillow==7.2.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
svgwrite==1.4
Tree==0.2.4
urllib3==1.25.10
w3lib==1.22.0
